migrate resolver service to new web.rest framework
