+++ python/nistoar/pdr/public/resolve/handlers/__init__.py
@@ -3,4 +3,4 @@ A subpackage providing ID resolution handlers.  Handlers implement the Handler c
 """
 from .pdrid import PDRIDHandler
 from .aip import AIPHandler
-from .base import Ready
+from .ready import ResolverReady
